# Conflicts:
#	app/src/main/java/com/mnassa/di/MnassaModules.kt
#	app/src/main/java/com/mnassa/screen/comments/CommentsWrapperController.kt
#	app/src/main/java/com/mnassa/screen/posts/PostDetailsFactory.kt
#	data/src/main/res/values/android_dictionary.xml

+++ app/src/main/java/com/mnassa/widget/coins/InOutComeView.kt
@@ -76,7 +76,7 @@ class InOutComeView : FrameLayout {
     }
 
 
-    fun showView(count: Int, name: String?) {
+    fun showView(count: Long, name: String?) {
         val from = StringBuilder()//"You just receive 50 points from ${name?:"Admin"}"
         paint.isAntiAlias = true
         if (count < 0) {
